ai-agent-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
# AI and LangChain dependencies
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
tiktoken==0.5.2
# Authentication dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
cryptography==41.0.7
# Security and validation
bleach==6.1.0
validators==0.22.0
# MCP (Model Context Protocol) dependencies
mcp==1.0.0
mcp-server==0.1.0
# Additional tool dependencies
requests==2.31.0
aiohttp==3.9.1
